requirements.txt
python-dotenv
openai
flask
flask_sqlalchemy
flask_socketio
pymysql
cryptography
config
instagrapi
Pillow
langchain